Extract setupBackToHotkeys.js
